+++ src/items.py
@@ -1,0 +1,46 @@
+Blocks = {
+0: {"TileName": "air", "Items": [{"Name": "Air"}]},
+1: {"TileName": "stone", "Items": [ 
+	{"Name": "Stone"},
+	{"Name": "Granite"},
+	{"Name": "Polished Granite"},
+	{"Name": "Diorite"},
+	{"Name": "Andesite"},
+	{"Name": "Polished Andesite"}
+]},
+2: {"TileName": "dirt", "Items": [
+	{"Name": "Dirt"},
+	{"Name": "Coarse Dirt"},
+	{"Name": "Podzol"}
+]},
+3: {"TileName": "grass", "Items": [{"Name": "Grass"}]},
+4: {"TileName": "cobblestone", "Items": [{"Name": "Cobblestone"}]},
+5: {"TileName": "planks", "Items": [
+	{"Name": "Oak Wood Plank"}, 
+	{"Name": "Spruce Wood Plank"}, 
+	{"Name": "Birch Wood Plank"},
+	{"Name": "Jungle Wood Plank"},
+	{"Name": "Acacia Wood Plank"},
+	{"Name": "Dark Oak Wood Plank"}
+]},
+6: {"TileName": "sapling", "Items": [
+	{"Name": "Oak Sapling"},
+	{"Name": "Spruce Sapling"},
+	{"Name": "Birch Sapling"},
+	{"Name": "Jungle Sapling"},
+	{"Name": "Acacia Sapling"},
+	{"Name": "Dark Oak Sapling"}
+]},
+7: {"TileName": "bedrock", "Items": [{"Name": "Bedrock"}]},
+8: {"TileName": "flowing_water", "Items": [{"Name": "Flowing Water"}]},
+9: {"TileName": "water", "Items": [{"Name": "Still Water"}]},
+10: {"TileName": "flowing_lava", "Items": [{"Name": "Flowing Lava"}]},
+11: {"TileName": "lava", "Items": [{"Name": "Still Lava"}]},
+12: {"TileName": "sand", "Items": [
+	{"Name": "Sand"},
+	{"Name": "Red Sand"}
+]},
+13: {"TileName": "gravel", "Items": [{"Name": "Gravel"}]},
+14: {"TileName": "gold_ore", "Items": [{"Name": "Gold Ore"}]},
+}
+# 0: {"TileName": "air", [{"Name": "Air"}]},